% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_fips.R
\name{harmonize_fips}
\alias{harmonize_fips}
\title{Harmonize FIPS codes across time periods}
\usage{
harmonize_fips(
  df,
  fips_col,
  year,
  target_year = NULL,
  weight_vars = NULL,
  agg_vars = NULL,
  group_vars = NULL,
  crosswalk = NULL,
  weight_type = "provided",
  handle_special_cases = TRUE,
  output_format = "standard",
  na.rm = T
)
}
\arguments{
\item{df}{A dataframe containing FIPS codes}

\item{fips_col}{The name of the column containing FIPS codes}

\item{year}{The year the FIPS codes in the data are from}

\item{target_year}{The year to convert FIPS codes to (defaults to most recent)}

\item{weight_vars}{Optional vector of column names containing values to be weighted during aggregation}

\item{agg_vars}{Optional vector of column names to be aggregated (using sum for numeric, first for non-numeric)}

\item{group_vars}{Optional vector of column names to group by (besides FIPS) when aggregating}

\item{crosswalk}{A dataframe containing FIPS code changes (must be provided)}

\item{weight_type}{How to handle weights: "provided" (use weights in crosswalk) or "equal" (split equally)}

\item{handle_special_cases}{Logical; whether to apply special handling for cases like CT COGs (default TRUE)}

\item{output_format}{The format for the output FIPS column, either "standard" or "nhgis".}

\item{na.rm}{Whether to remove NAs in weighting, by default "TRUE"}
}
\value{
A dataframe with harmonized FIPS codes and appropriately weighted/aggregated values
}
\description{
Harmonize FIPS codes across time periods
}
