% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_fips.R
\name{reformat_fips}
\alias{reformat_fips}
\title{Reformat FIPS codes}
\usage{
reformat_fips(fips_vector, output_format = "standard")
}
\arguments{
\item{fips_vector}{A vector of FIPS codes to reformat.}

\item{output_format}{The desired output format. Can be "standard" (for 5-digit FIPS)
or "nhgis". Defaults to "standard".}
}
\value{
A character vector of reformatted FIPS codes.
}
\description{
This function reformats a vector of FIPS codes to a standard 5-digit format
or the NHGIS format. It can handle various input formats, including numeric
FIPS, character FIPS with or without leading zeros, and NHGIS formatted FIPS codes.
}
\examples{
reformat_fips(c(1001, "01001", "G010010"))
reformat_fips(c(1001, "01001", "G010010"), output_format = "nhgis")
}
